requirements.txt
fastapi==0.111.0
uvicorn==0.21.1
python-socketio==5.11.2
fastapi-socketio==0.0.10
loguru==0.7.2